RedBlackTree: eager rebalance after insert
Change internal rebalance timing.  Do convert (b (r r)) -> (r (b b)) to
make balance of tree be perfect after insert.  Before this commit, the
above convert run at the next insert.

It includes related 4 updates for the test.  2 tests for delete indicate
that the balance of the tree is better than before.  Other 2 tests for
insert indicate that the balance of the tree is better than before at
the first insert, and the balance is the same as before at the second
insert.